feat: api 코드수정, 브라우저 css 제거
