+++ src/app/_hooks/useCreateEvent.ts
@@ -1,13 +1,15 @@
-import { apiClient } from "@/utils/axios";
-import { EventCreateFormData } from "@/utils/event-create-formdata";
 import { useMutation } from "@tanstack/react-query";
+import { apiClient } from "@/utils/axios";
 import { EventFormValues } from "app/_types/event";
+import { createEventFormData } from "@/utils/create-event-formdata";
 
 export const useCreateEvent = () => {
   return useMutation({
     mutationFn: async (payload: EventFormValues) => {
-      const formData = EventCreateFormData(payload);
+      const formData = createEventFormData(payload);
+
       const res = await apiClient.post("/api/events", formData);
+
       return res.data;
     },
   });
