+++ src/app/_styles/globals.css
@@ -191,3 +191,18 @@ body {
   box-shadow: 0px 0px 80px 1px rgba(0, 0, 0, 0.2);
   backdrop-filter: blur(22.5px);
 }
+
+/* 자동입력 시 브라우저 css 배경색 제거 */
+input:-webkit-autofill,
+input:-webkit-autofill:hover,
+input:-webkit-autofill:focus,
+textarea:-webkit-autofill,
+textarea:-webkit-autofill:hover,
+textarea:-webkit-autofill:focus,
+select:-webkit-autofill {
+  -webkit-box-shadow: 0 0 0px 1000px transparent inset !important;
+  box-shadow: 0 0 0px 1000px transparent inset !important;
+  transition: background-color 5000s ease-in-out 0s !important;
+  background-color: transparent !important;
+  color: inherit !important;
+}
